foreign keys adjustments

+++ database/migrations/2020_10_16_082603_create_apartment_user_table.php
@@ -16,8 +16,8 @@ class CreateApartmentUserTable extends Migration
         Schema::create('apartment_user', function (Blueprint $table) {
             $table->id();
 
-            $table -> integer('user_id') -> unsigned();
-            $table -> integer('apartment_id') -> unsigned();
+            $table -> bigInteger('user_id') -> unsigned();
+            $table -> bigInteger('apartment_id') -> unsigned();
             
             $table->timestamps();
         });
